Scrolling doesn't work with iPhone 4
Fixes #2. The Autoresize Mask is set which is resizing the control.
Clipping of the subview is not enabled so the dial is still being
drawn on screen.

Added clipping of the subview so now you can see the control show you
if it is resized. Also disable the Autoresize mask to make sure the
control doesn't resize in the demo project.